requirements.txt
google-cloud-pubsub==1.6.0
tweepy==3.8.0